Use correct type size for port elements
The ports which are used for TCP pings are stored
in a GArray. TCP ports are of size uint16_t.
Now the array elements are all of size uint16_t.
When ports are added to the array they are first
cast to the appropriate type. The port lists
was already validated, so there are no values which
are bigger than uint16_t in it.

+++ boreas/alivedetection.c
@@ -339,7 +339,8 @@ alive_detection_init (gvm_hosts_t *hosts, alive_test_t alive_test)
        * again. */
       port_list = prefs_get ("port_range");
     }
-  scanner.ports = g_array_new (FALSE, TRUE, sizeof (int));
+  /* Use uint16_t for port array elements. tcphdr port type is uint16_t. */
+  scanner.ports = g_array_new (FALSE, TRUE, sizeof (uint16_t));
   if (port_list)
     portranges_array = port_range_ranges (port_list);
   else
